# openapi/activation.openapi.yaml
openapi: 3.0.3
info:
  title: RTP Activation API.
  version: 1.0.0
  description: |
    API to handle RTP activations initiated by the Payer's RTP Service Provider.
  contact:
    name: PagoPA S.p.A.
    email: rtp@pagopa.it
servers:
  - description: Development/Test
    url: localhost:8080
    x-internal: true
...
paths:
  /activations:
    post:
      operationId: activate
      summary: RTP activation initiated by the Payer's RTP Service Provider.
      description: |
        The operation is used by Payer's RTP Service Provider to enable the
        Payee's RTP Service Provider to send RTP messages to the Payer's RTP
        Service Provider.
        
        When the operation is used by not-admin subject, the system verifies
        that the Payer's RTP Service Provider ID matches the subject claim of
        the access token. `403 Forbidden` is returned on mismatch.
        
        `409 Conflict` is returned if an activation with the same Payer's ID
        already exists.
      tags: [create]
      security:
        - oAuth2: [admin_rtp_activations, write_rtp_activations]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
      requestBody:
        $ref: '#/components/requestBodies/CreateOrUpdateActivation'
      responses:
        "201":
          #description: Created.
          $ref: '#/components/responses/CreateActivation'
        "400":
          #description: Bad request.
          $ref: '#/components/responses/Error'
        "401":
          #description: Wrong credentials.
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/ForbiddenError'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/GenericError'
        "409":
          #description: Conflict.
          $ref: '#/components/responses/ConflictError'
        "415":
          #description: Unsupported media type. Did you provide application/json?
          $ref: '#/components/responses/GenericError'
        "429":
          #description: Too many request.
          $ref: '#/components/responses/GenericError'
        "500":
          #description: Server error.
          $ref: '#/components/responses/GenericError'
        default:
          #description: Unexpected error.
          $ref: '#/components/responses/GenericError'

    get:
      operationId: getActivations
      summary: Returns RTP activations.
      description: |
        The operation returns all the RTP activations stored by the system.
        
        When the operation is used by not-admin subject, the system returns
        the RTP activations which have the Payer's RTP Service Provider ID that
        matches the subject claim of the access token.
      tags: [read]
      security:
        - oAuth2: [admin_rtp_activations, read_rtp_activations]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        "200":
          #description: Ok.
          $ref: '#/components/responses/PageOfActivations'
        "400":
          #description: Bad request.
          $ref: '#/components/responses/Error'
        "401":
          #description: Access token is missing or invalid.
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden.
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request.
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error.
          $ref: '#/components/responses/Error'
        default:
          #description: Unexpected error.
          $ref: '#/components/responses/Error'
  /activations/{activationId}:
    get:
      operationId: getActivation
      summary: Returns a RTP activation.
      description: |
        The operation finds a RTP activation by its ID.
        
        When the operation is used by not-admin subject, the system returns
        the RTP activation only if its Payer's RTP Service Provider ID matches
        the subject claim of the access token, otherwise `404 Not Found` is
        returned.
      tags: [read]
      security:
        - oAuth2: [admin_rtp_activations, read_rtp_activations]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/ActivationId'
      responses:
        "200":
          #description: Found.
          $ref: '#/components/responses/Activation'
        "400":
          #description: Bad request.
          $ref: '#/components/responses/Error'
        "401":
          #description: Access token is missing or invalid.
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden.
          $ref: '#/components/responses/Error'
        "404":
          #description: Not found.
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request.
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error.
          $ref: '#/components/responses/Error'
        default:
          #description: Unexpected error.
          $ref: '#/components/responses/Error'

    put:
      operationId: updateActivation
      summary: Updates a RTP activation.
      description: |
        The operation updates a RTP activation searching it by its ID.
        
        When the operation is used by not-admin subject:
          - the system returns `404 Not Found` if the Payer's RTP Service
            Provider ID of the activation doesn't match the subject claim of
            the access token;
          - the system returns `403 Forbidden` if the provided value of Payer's
            RTP Service doesn't match the subject claim of the access token.
      tags: [update]
      security:
        - oAuth2: [admin_rtp_activations, write_rtp_activations]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/ActivationId'
      requestBody:
        $ref: '#/components/requestBodies/CreateOrUpdateActivation'
      responses:
        "204":
          #description: No content
          $ref: '#/components/responses/NoContent'
        "400":
          #description: Bad request
          $ref: '#/components/responses/Error'
        "401":
          #description: Wrong credentials
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/Error'
        "404":
          #description: Not found
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "415":
          #description: Unsupported media type. Did you provide application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error
          $ref: '#/components/responses/Error'
        default:
          #description: Unexpected error
          $ref: '#/components/responses/Error'

    delete:
      operationId: deleteActivation
      summary: Deletes a RTP activation.
      description: |
        The operation deletes a RTP activation searching it by its ID.
        
        When the operation is used by not-admin subject, the system returns
        `404 Not Found` if the Payer's RTP Service Provider ID of the activation
        doesn't match the subject claim of the access token.
      tags: [delete]
      security:
        - oAuth2: [admin_rtp_activations, write_rtp_activations]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/ActivationId'
      responses:
        "204":
          #description: No content
          $ref: '#/components/responses/NoContent'
        "400":
          #description: Bad request
          $ref: '#/components/responses/Error'
        "401":
          #description: Access token is missing or invalid
          $ref: '#/components/responses/Error'
        "403":
          #description: Forbidden
          $ref: '#/components/responses/Error'
        "404":
          #description: Not found
          $ref: '#/components/responses/Error'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/Error'
        "429":
          #description: Too many request
          $ref: '#/components/responses/Error'
        "500":
          #description: Server error
          $ref: '#/components/responses/Error'
        default:
          #description: Unexpected error
          $ref: '#/components/responses/Error'
  /activations/payer:
    get:
      operationId: findActivationByPayerId
      summary: Finds a RTP activation by Payer ID.
      description: |
        The operation finds RTP activation by Payer ID.
        
        When the operation is used by not-admin subject, the system returns
        the RTP activation with the Payer's RTP Service Provider ID that
        matches the subject claim of the access token.
      tags: [read]
      security:
        - oAuth2: [admin_rtp_activations, read_rtp_activations]
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/Version'
        - $ref: '#/components/parameters/PayerId'
      responses:
        "200":
          #description: Ok.
          $ref: '#/components/responses/Activation'
        "400":
          #description: Bad request.
          $ref: '#/components/responses/Error'
        "401":
          #description: Access token is missing or invalid.
          $ref: '#/components/responses/UnauthorizedError'
        "403":
          #description: Forbidden.
          $ref: '#/components/responses/ForbiddenError'
        "406":
          #description: Not acceptable. Did you require application/json?
          $ref: '#/components/responses/GenericError'
        "429":
          #description: Too many request.
          $ref: '#/components/responses/GenericError'
        "500":
          #description: Server error.
          $ref: '#/components/responses/GenericError'
        default:
          #description: Unexpected error.
          $ref: '#/components/responses/GenericError'
components:
...
  schemas:
...
    AccessControlAllowOrigin:
      description: |
        Indicates whether the response can be shared with requesting code from
        the given origin.
      type: string
      pattern: "^[ -~]{1,2048}$"
      minLength: 1
      maxLength: 2048

    # --------------------------------------------------------------------------
    # Basic types for rate limit handling.
    # --------------------------------------------------------------------------
    RateLimitLimit:
      description: The number of allowed requests in the current period.
      type: integer
      format: int32
      minimum: 1
      maximum: 240
    RateLimitReset:
      description: The number of seconds left in the current period.
      type: integer
      format: int32
      minimum: 1
      maximum: 60
    RetryAfter:
      description: |
        The number of seconds to wait before allowing a follow-up request.
      type: integer
      format: int32
      minimum: 1
      maximum: 240

    # --------------------------------------------------------------------------
    # Basic types for paging.
    # --------------------------------------------------------------------------
    PageNumber:
      description: Number of the page.
      type: integer
      format: int32
      minimum: 0
      maximum: 2147483647
      example: 1
    PageSize:
      description: Size of the page.
      type: integer
      format: int32
      minimum: 1
      maximum: 128
      example: 20
...
    ErrorCode:
      description: Error code.
      type: string
      pattern: "^[A-F0-9]{9}$"
      minLength: 9
      maxLength: 9
      example: "01000000F"
...
    ActivationLocation:
      description: URL of the RTP activation resource.
      type: string
      format: uri
      pattern: "^[ -~]{1,2048}$"
      minLength: 1
      maxLength: 2048
      example: "https://rtp.cstar.pagopa.it/activations/d0d654e6-97da-4848-b568-99fedccb642b"
    RequestId:
      description: Identifier of the request.
      type: string
      format: uuid
      pattern: "^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$"
      minLength: 36
      maxLength: 36
      example: "bd615b4a-066d-443e-8dd2-a28a39931fef"
    Version:
      description: Version of the required API.
      type: string
      pattern: "^[ -~]{1,64}$"
      minLength: 1
      maxLength: 64
      example: "v1"

    # --------------------------------------------------------------------------
    # Domain specific basic types.
    # --------------------------------------------------------------------------
...
    Errors:
      description: List of errors.
      type: object
      additionalProperties: false
      properties:
        errors:
          type: array
          minItems: 1
          maxItems: 32
          items:
            $ref: '#/components/schemas/Error'
      required:
        - errors
      example:
        errors:
          - statusCode: "01000000F"
            description: "Wrong party identifier"
    SyntheticError:
      description: Synthetic error details.
      type: object
      additionalProperties: false
      properties:
        statusCode:
          $ref: '#/components/schemas/ErrorCode'
        description:
          $ref: '#/components/schemas/ErrorDescription'
      required:
        - code
        - description
      example:
        statusCode: "01000000F"
        description: "Invalid JWT."
    GenericError:
      description: Generic error.
      type: object
      additionalProperties: false
      properties:
        timestamp:
          type: string
          format: date-time
          description: The time the error occurred.
        path:
          type: string
          description: The API path where the error occurred.
        status:
          type: integer
          description: HTTP status code of the error.
        statusCode:
          type: integer
          description: HTTP status code of the error.
        error:
          type: string
          description: Description of the error.
        message:
          type: string
          description: Description of the error.
        requestId:
          type: string
          description: Unique identifier for the request.
      example:
        timestamp: "2024-12-31T09:54:01.763+00:00"
        path: "/activations"
        status: 415
        error: "Unsupported Media Type"
        requestId: "3fb00d0f-416"

    # ------------------------------------------------------
    # Domain specific complex types.
    # ------------------------------------------------------
...
    ActivationReq:
      description: |
        Data of a RTP activation.
      type: object
      additionalProperties: true # It's extended by another object.
      properties:
        payer:
          $ref: '#/components/schemas/Payer'
      required:
        - payer
      example:
        payer:
          fiscalCode: "RSSMRA85T10A562S"
          rtpSpId: "12345678911"
...
  requestBodies:
    CreateOrUpdateActivation:
      description: Request to create or update a RTP activation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ActivationReq'

  # ============================================================================
  # Parameters.
  # ============================================================================
  parameters:
    ActivationId:
      name: activationId
      in: path
      description: Identifier of the RTP activation resource.
      required: true
      schema:
        $ref: '#/components/schemas/ActivationId'
    PageNumber:
      name: page
      in: query
      description: Number of the requested page of data.
      required: true
      schema:
        $ref: '#/components/schemas/PageNumber'
    PageSize:
      name: size
      in: query
      description: Size of the requested page of data.
      required: true
      schema:
        $ref: '#/components/schemas/PageSize'
    PayerId:
      name: PayerId
      in: header
      description: Identifier of the Payer.
      required: true
      schema:
        $ref: '#/components/schemas/FiscalCode'
    RequestId:
      name: RequestId
      in: header
      description: Identifier of the request.
      required: true
      schema:
        $ref: '#/components/schemas/RequestId'
    Version:
      name: Version
      in: header
      description: Version of the required API.
      required: false
      schema:
        $ref: '#/components/schemas/Version'

  # ============================================================================
  # Responses
  # ============================================================================
  responses:
    Activation:
      description: Response returned when RTP activation data is requested.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Activation'
    CreateActivation:
      description: Response returned when a RTP activation is requested.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Location:
          description: URL of the activation resource.
          required: true
          schema:
            $ref: '#/components/schemas/ActivationLocation'
    UnauthorizedError:
      description: Response returned when the credentials of the user are missing or invalid.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Retry-After:
          description: |
            The number of seconds to wait before allowing a follow-up request.
          required: false
          schema:
            $ref: '#/components/schemas/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SyntheticError'
        text/*:
          schema:
            type: string
            pattern: "^[ -~]{0,65535}$"
            maxLength: 65535
    ForbiddenError:
      description: Response returned when the user does not have the right permissions.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
    ConflictError:
      description: Response returned when a RTP payer is already active.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
    Error:
      description: Error response detailing constraint validations.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Retry-After:
          description: |
            The number of seconds to wait before allowing a follow-up request.
          required: false
          schema:
            $ref: '#/components/schemas/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errors'
        text/*:
          schema:
            type: string
            pattern: "^[ -~]{0,65535}$"
            maxLength: 65535
    GenericError:
      description: Generic Error response.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
        Retry-After:
          description: |
            The number of seconds to wait before allowing a follow-up request.
          required: false
          schema:
            $ref: '#/components/schemas/RetryAfter'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericError'
        text/*:
          schema:
            type: string
            pattern: "^[ -~]{0,65535}$"
            maxLength: 65535
    NoContent:
      description: No content response.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
    PageOfActivations:
      description: Response to the request to get RTP activations.
      headers:
        Access-Control-Allow-Origin:
          description: |
            Indicates whether the response can be shared with requesting code
            from the given origin.
          required: false
          schema:
            $ref: '#/components/schemas/AccessControlAllowOrigin'
        RateLimit-Limit:
          description: The number of allowed requests in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitLimit'
        RateLimit-Reset:
          description: The number of seconds left in the current period.
          required: false
          schema:
            $ref: '#/components/schemas/RateLimitReset'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PageOfActivations'

  # ============================================================================
  # Security schemes.
  # ============================================================================
  securitySchemes:
    oAuth2:
      description: |
        A bearer token in the format of a JWS and conforms to the specifications
        included in RFC8725.
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: /token
          refreshUrl: /token
          scopes:
            admin_rtp_activations: Admin RPT activation.
            write_rtp_activations: Create, update or delete RTP activation.
            read_rtp_activations: Read RTP activation.